Move platform-specific impls of EdgeDaemon to their own files
